refactor(FeedHeart) #420: 인터페이스 다형성 추가
