+++ be/src/main/java/com/foodymoody/be/feed_heart/infra/persistence/jpa/FeedHeartJpaRepository.java
@@ -1,4 +1,4 @@
-package com.foodymoody.be.feed_heart.domain;
+package com.foodymoody.be.feed_heart.infra.persistence.jpa;
 
 import com.foodymoody.be.common.util.ids.FeedHeartId;
 import com.foodymoody.be.common.util.ids.FeedId;
@@ -6,9 +6,7 @@ import com.foodymoody.be.common.util.ids.MemberId;
 import com.foodymoody.be.feed_heart.domain.entity.FeedHeart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedHeartRepository {
-
-    FeedHeart save(FeedHeart feedHeart);
+public interface FeedHeartJpaRepository extends JpaRepository<FeedHeart, FeedHeartId> {
 
     boolean existsHeartByMemberIdAndFeedId(MemberId memberId, FeedId feedId);
